+++ app/javascript/components/splitwisePayment/SplitwisePaymentConfirm.vue
@@ -28,7 +28,7 @@
 
 <script>
 import { mapState } from 'vuex';
-import textField from '../components/TextField';
+import textField from '../TextField';
 
 export default {
   name: 'SplitwisePaymentConfirm',
